fix(clients): float usage type, paced signal-quality fan-out, demote per-network logs
- F-112: widen NetworkClient.usage and ClientData.usage to dict[str, float] so a
  fractional KB value no longer aborts the whole network's client collection
  (verified against live API data)
- F-060: gate + cap + rate-limit the per-client wireless signal-quality fan-out
- F-171: demote per-network client log lines to debug, keep one INFO summary
- also defines the F-029 device domain models (wired into collectors next commit)

Closes #492
Closes #493
Closes #494

+++ src/meraki_dashboard_exporter/core/api_models.py
@@ -279,7 +279,10 @@ class NetworkClient(BaseModel):
     ssid: str | None = None
     vlan: str | None = None  # Can be string or None in API
     switchport: str | None = None
-    usage: dict[str, int] | None = None
+    # Live API returns FLOAT kilobyte values (e.g. {"sent": 225.6, "recv": 852.5});
+    # typing this as int made model_validate raise on fractional usage and drop the
+    # whole network's client collection for that cycle. Floats accept ints too.
+    usage: dict[str, float] | None = None
     status: Literal["Online", "Offline"] | str = "Offline"
     notes: str | None = None
     groupPolicy8021x: str | None = None
